+++ hoot-services/src/test/java/hoot/services/controllers/osm/ChangesetResourceCloseTest.java
@@ -868,11 +868,11 @@ public class ChangesetResourceCloseTest extends OsmResourceTestAbstract
         CurrentNodes nodeRecord = (CurrentNodes)nodes.get(nodeIdsArr[0]);
         Assert.assertEquals(new Long(changesetId), nodeRecord.getChangesetId());
         Assert.assertEquals(
-          new Integer((int)(DbUtils.toDbCoordPrecision(updatedBounds.getMinLat()) *
+          new Long((long)(DbUtils.toDbCoordPrecision(updatedBounds.getMinLat()) *
             GeoUtils.GEO_RECORD_SCALE)),
           nodeRecord.getLatitude());
         Assert.assertEquals(
-          new Integer((int)(DbUtils.toDbCoordPrecision(updatedBounds.getMinLon()) *
+          new Long((long)(DbUtils.toDbCoordPrecision(updatedBounds.getMinLon()) *
             GeoUtils.GEO_RECORD_SCALE)),
           nodeRecord.getLongitude());
         Assert.assertEquals(nodeIdsArr[0], nodeRecord.getId());
@@ -894,11 +894,11 @@ public class ChangesetResourceCloseTest extends OsmResourceTestAbstract
         nodeRecord = (CurrentNodes)nodes.get(nodeIdsArr[1]);
         Assert.assertEquals(new Long(changesetId), nodeRecord.getChangesetId());
         Assert.assertEquals(
-          new Integer((int)(DbUtils.toDbCoordPrecision(updatedBounds.getMinLat()) *
+          new Long((long)(DbUtils.toDbCoordPrecision(updatedBounds.getMinLat()) *
             GeoUtils.GEO_RECORD_SCALE)),
           nodeRecord.getLatitude());
         Assert.assertEquals(
-          new Integer((int)(DbUtils.toDbCoordPrecision(originalBounds.getMaxLon()) *
+          new Long((long)(DbUtils.toDbCoordPrecision(originalBounds.getMaxLon()) *
             GeoUtils.GEO_RECORD_SCALE)),
           nodeRecord.getLongitude());
         Assert.assertEquals(nodeIdsArr[1], nodeRecord.getId());
