+++ hoot-services/src/test/java/hoot/services/controllers/osm/ChangesetResourceUploadCreateTest.java
@@ -480,11 +480,11 @@ public class ChangesetResourceUploadCreateTest extends OsmResourceTestAbstract
         CurrentNodes nodeRecord = (CurrentNodes)nodes.get(nodeIdsArr[0]);
         Assert.assertEquals(new Long(changesetId), nodeRecord.getChangesetId());
         Assert.assertEquals(
-          new Integer((int)(DbUtils.toDbCoordPrecision(originalBounds.getMinLat()) *
+          new Long((long)(DbUtils.toDbCoordPrecision(originalBounds.getMinLat()) *
             GeoUtils.GEO_RECORD_SCALE)),
           nodeRecord.getLatitude());
         Assert.assertEquals(
-          new Integer((int)(DbUtils.toDbCoordPrecision(originalBounds.getMinLon()) *
+          new Long((long)(DbUtils.toDbCoordPrecision(originalBounds.getMinLon()) *
             GeoUtils.GEO_RECORD_SCALE)),
           nodeRecord.getLongitude());
         Assert.assertEquals(nodeIdsArr[0], nodeRecord.getId());
@@ -500,11 +500,11 @@ public class ChangesetResourceUploadCreateTest extends OsmResourceTestAbstract
         nodeRecord = (CurrentNodes)nodes.get(nodeIdsArr[1]);
         Assert.assertEquals(new Long(changesetId), nodeRecord.getChangesetId());
         Assert.assertEquals(
-          new Integer((int)(DbUtils.toDbCoordPrecision(originalBounds.getMaxLat()) *
+          new Long((long)(DbUtils.toDbCoordPrecision(originalBounds.getMaxLat()) *
             GeoUtils.GEO_RECORD_SCALE)),
           nodeRecord.getLatitude());
         Assert.assertEquals(
-          new Integer((int)(DbUtils.toDbCoordPrecision(originalBounds.getMaxLon()) *
+          new Long((long)(DbUtils.toDbCoordPrecision(originalBounds.getMaxLon()) *
             GeoUtils.GEO_RECORD_SCALE)),
           nodeRecord.getLongitude());
         Assert.assertEquals(nodeIdsArr[1], nodeRecord.getId());
@@ -520,11 +520,11 @@ public class ChangesetResourceUploadCreateTest extends OsmResourceTestAbstract
         nodeRecord = (CurrentNodes)nodes.get(nodeIdsArr[2]);
         Assert.assertEquals(new Long(changesetId), nodeRecord.getChangesetId());
         Assert.assertEquals(
-          new Integer((int)(DbUtils.toDbCoordPrecision((originalBounds.getMinLat() - .001)) *
+          new Long((long)(DbUtils.toDbCoordPrecision((originalBounds.getMinLat() - .001)) *
             GeoUtils.GEO_RECORD_SCALE)),
           nodeRecord.getLatitude());
         Assert.assertEquals(
-          new Integer((int)(DbUtils.toDbCoordPrecision((originalBounds.getMinLon() - .001)) *
+          new Long((long)(DbUtils.toDbCoordPrecision((originalBounds.getMinLon() - .001)) *
             GeoUtils.GEO_RECORD_SCALE)),
           nodeRecord.getLongitude());
         Assert.assertEquals(nodeIdsArr[2], nodeRecord.getId());
